+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Dev: Improve migration tests. (#290)
 - Dev: Improve Twitter tests. (#293)
 - Dev: Improve SevenTV tests. (#294)
+- Dev: Improve FrankerFaceZ tests. (#295)
 
 ## 1.2.3